Add projection preconditioner based on thyra
